Various testing updates
