+++ Tests/Tests.csproj
@@ -50,8 +50,11 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="BoardTests.cs" />
+    <Compile Include="Cutting.cs" />
+    <Compile Include="Main.cs" />
+    <Compile Include="Creation.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="VerySlow.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SudokuSharp.csproj">
